[538387] Clean HTTP and add function for adding configurations
The sockets are closed in a better way and some cleanup was also made.
A new cmake function allows custom configurations that are added to the
forte_config.h generated file

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=538387

+++ src/core/devexec.h
@@ -57,7 +57,7 @@ public:
 
     CTimerHandler& getTimer() const;
 
-    CExternalEventHandler* getHandler(unsigned int paIdentifer) const;
+    CExternalEventHandler* getHandler(size_t paIdentifer) const;
 
 protected:
 private:
